+++ lib/common/core/blocs/radio/radio_event.dart
@@ -1,0 +1,10 @@
+part of 'radio_bloc.dart';
+
+@immutable
+abstract class RadioEvent {}
+
+class SwitchRadioEvent<V> extends RadioEvent {
+  final V index;
+
+  SwitchRadioEvent(this.index);
+}