+++ CHANGELOG.md
@@ -1,3 +1,19 @@
+## 0.1.6
+
+Add documentation for take logic.
+
+## 0.1.5
+
+Add documentation for valid logic.
+
+## 0.1.4
+
+Add documentation for check logic.
+
+## 0.1.3
+
+Add radio logic.
+
 ## 0.1.2
 
 Add examples for TakeLogic, ValidLogic, CheckLogic.